+++ src/server/maupod/cmd/restapi/pkg/api/genres.go
